+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :bug:`-` fix Well.set_properties() so that it completely overwrites the existing properties dict
 * :bug:`-` respect incubate conditions where uncovered=True
 * :bug:`-` prevent invalid incubate parameters in :ref:`protocol-absorbance`
 * :bug:`-` allow incubation of containers at ambient without covers